# specs/001-specify-scripts-bash/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: FT Backend Core Services API
  version: 0.1.0
servers:
  - url: https://localhost:3000
    description: Development via Docker Compose
paths:
  /auth/register:
    post:
      summary: Register via email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                displayName:
                  type: string
              required: [email, password, displayName]
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
  /auth/login:
    post:
      summary: Login with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required: [email, password]
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
  /auth/token/refresh:
    post:
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
              required: [refreshToken]
      responses:
        '200':
          description: Tokens rotated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
  /auth/logout:
    post:
      summary: Revoke active session
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Session revoked
...
  /auth/42/callback:
    get:
      summary: Handle 42 OAuth callback
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OAuth success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
  /auth/me:
    get:
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
...
  /chat/channels:
    get:
      summary: List chat channels
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Channel list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatChannel'
    post:
      summary: Create channel
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatChannelCreate'
      responses:
        '201':
          description: Channel created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatChannel'
...
  /chat/dm/{userId}:
    post:
      summary: Send direct message
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DMCreate'
      responses:
        '201':
          description: DM stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessage'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    AuthTokens:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
      required: [accessToken, refreshToken, expiresIn]
    UserProfile:
      type: object
      properties:
        id:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
        bio:
          type: string
      required: [id, displayName]
...
    ChatChannel:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        visibility:
          type: string
      required: [id, slug, title]
    ChatChannelCreate:
      type: object
      properties:
        title:
          type: string
        visibility:
          type: string
          enum: [public, private]
      required: [title]
    ChatMessage:
      type: object
      properties:
        id:
          type: string
        channelId:
          type: string
        senderId:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, channelId, senderId, content, createdAt]
...
    DMCreate:
      type: object
      properties:
        content:
          type: string
      required: [content]
...